feat: 试剂成本、委外成本计算

+++ yudao-module-jl/yudao-module-jl-biz/src/main/java/cn/iocoder/yudao/module/jl/controller/admin/project/vo/ProjectCostStatsRespVO.java
@@ -22,7 +22,7 @@ public class ProjectCostStatsRespVO {
 
 
     @Schema(description = "物资成本")
-    private Long supplyCost;
+    private BigDecimal supplyCost;
 
     @Schema(description = "采购成本")
     private Long procurementCost;
